api/todo/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["jawwad-ali <alijawwad001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.0.2"
httpx = "^0.27.0"
sqlmodel = "^0.0.16"
aiokafka = "^0.10.0"
openai = "^1.30.5"
google-generativeai = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"